Some 2nd pass effects, 3rd pass offscreen in use

+++ GDP_Feeney_201718/assimp/cSimpleAssimpSkinnedMeshLoader_OneMesh.cpp
@@ -107,7 +107,7 @@ float cSimpleAssimpSkinnedMesh::FindAnimationTotalTime(std::string animationName
 	}
 
 	// This is scaling the animation from 0 to 1
-	return itAnimation->second->mAnimations[0]->mDuration;
+	return (float)itAnimation->second->mAnimations[0]->mDuration;
 }
